Changed types for height variables

+++ 0x1D-binary_trees/16-binary_tree_is_perfect.c
@@ -8,8 +8,8 @@
  */
 size_t binary_tree_height(const binary_tree_t *tree)
 {
-	size_t height_l;
-	size_t height_r;
+	int height_l;
+	int height_r;
 
 	height_l = tree->left ? 1 + binary_tree_height(tree->left) : 0;
 	height_r = tree->right ? 1 + binary_tree_height(tree->right) : 0;
@@ -24,8 +24,8 @@ size_t binary_tree_height(const binary_tree_t *tree)
  */
 int binary_tree_is_perfect(const binary_tree_t *tree)
 {
-	size_t height_l;
-	size_t height_r;
+	int height_l;
+	int height_r;
 
 	if (!tree)
 		return (0);
